+++ unnamed_client/src/client/mod.rs
@@ -1,2 +1,0 @@
-//! ## Client
-pub mod client;