+++ unnamed_editor/src/main.rs
@@ -1,9 +1,0 @@
-//! ## Editor Entry-Point
-mod editor;
-
-use editor::editor::Editor;
-
-fn main() {
-  let mut editor = Editor::new("UnnamedEditor".to_string());
-  editor.start();
-}