+++ unnamed_engine/src/event/mod.rs
@@ -1,2 +1,0 @@
-//! ## Event
-pub mod event;