+++ unnamed_engine/src/gui/mod.rs
@@ -1,3 +1,0 @@
-//! ## GUI
-pub mod egui_renderer;
-pub mod gui;