+++ unnamed_engine/src/input/mod.rs
@@ -1,4 +1,0 @@
-//! Input
-pub mod input_manager;
-pub mod keyboard;
-pub mod mouse;