+++ unnamed_engine/src/lib.rs
@@ -1,10 +1,1 @@
-//! ## UnnamedEngine
 pub mod core;
-pub mod event;
-pub mod gui;
-pub mod input;
-pub mod networking;
-pub mod renderer;
-pub mod voxel;
-
-pub use log;