+++ unnamed_server/src/main.rs
@@ -1,9 +1,0 @@
-//! ## Server Entry-Point
-use server::server::Server;
-
-mod server;
-
-fn main() {
-  let mut server = Server::new("UnnamedServer".to_string());
-  server.start();
-}