+++ unnamed_server/src/server/mod.rs
@@ -1,2 +1,0 @@
-//! ## Server
-pub mod server;